# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Comet API",
    "version": "1.0.0",
    "description": "API documentation for Comet: The Speed of Light Comic Reader."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Local API"
    }
  ],
  "components": {
    "schemas": {
      "ReadingProgress": {
        "type": "object",
        "required": [
          "lastPage",
          "totalPages"
        ],
        "properties": {
          "lastPage": {
            "type": "integer",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "minimum": 1
          },
          "zoomLevel": {
            "type": "number",
            "default": 1
          },
          "readStatus": {
            "type": "string",
            "enum": [
              "UNREAD",
              "READING",
              "COMPLETED"
            ]
          },
          "timeDelta": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "ComicDTO": {
        "type": "object",
        "required": [
          "id",
          "title",
          "pageCount",
          "addedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string",
            "nullable": true
          },
          "coverUrl": {
            "type": "string",
            "nullable": true
          },
          "pageCount": {
            "type": "integer"
          },
          "year": {
            "type": "integer",
            "nullable": true
          },
          "series": {
            "type": "string",
            "nullable": true
          },
          "isFavorite": {
            "type": "boolean",
            "default": false
          },
          "rating": {
            "type": "integer",
            "minimum": 0,
            "maximum": 10,
            "default": 0
          },
          "syncStatus": {
            "type": "string",
            "enum": [
              "LOCAL",
              "PENDING",
              "SYNCED",
              "ERROR"
            ],
            "default": "LOCAL"
          },
          "filehash": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer"
          },
          "progress": {
            "$ref": "#/components/schemas/ReadingProgress"
          },
          "addedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PaginatedLibraryResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComicDTO"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/library": {
      "get": {
        "summary": "Get comic library",
        "tags": [
          "Library"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedLibraryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "get_library"
      }
    },
    "/library/upload": {
      "post": {
        "summary": "Upload comic file to S3",
        "tags": [
          "Library"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "filehash": {
                    "type": "string",
                    "minLength": 64,
                    "maxLength": 64
                  },
                  "extension": {
                    "type": "string",
                    "enum": [
                      "cbz",
                      "cbr"
                    ]
                  }
                },
                "required": [
                  "file",
                  "filehash",
                  "extension"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload successful"
          },
          "400": {
            "description": "Invalid format or missing data"
          },
          "413": {
            "description": "File too large"
          }
        },
        "operationId": "post_library_upload"
      }
    },
...
  }
}